Add a cookie-based server select dropdown.
Users may specify an array of servers in the $server variable at the top of the file. If $server is a string, or an array with a single value, it will function the same as before. But if $server contains multiple values, the user will be presented with a dropdown to switch between servers.